Added files via upload
Ok so this class takes in a file of cities and beats and returns a new list of cities with most of the blanks filled in. It also saves a list of "flags". These are blanks that had no recognizable beat and need to be filled in by hand. This list can be accessed with getFlags();