Added files via upload
Feature Being Added: Community Code Getter
Description: The community code getter takes in a file of reporting districts and returns an associated community code. It finds the community codes from a hash map created in the "MapSaver" class. It returns the community codes as an array list of strings like all the other classes so far.
Details: The community code getter class does not read the hash map from map saver. Since the hash map created in map saver only needs to be created once I have saved it into a .ser file. The code getter class draws directly from that file. For info on serialization and deserialization of hash maps see: "http://beginnersbook.com/2013/12/how-to-serialize-hashmap-in-java/" .