Added files via upload
Feature Being Added: APPS Formatter
Feature Description: Takes in an excel file with an assortment of numbers. Changes the 0's to NO's and any other number to YES. It makes these modifications directly to the excel sheet and the actual method returns one of two strings: The file is not closed, please close the spreadsheet and try again (in which case the excel file is unchanged) and File Formatted Successfully. 
Additional Comments: Going to try and go through some of the other files and change them to modify the excel sheet directly. I also included an example of the type of file this program will modify.