Added in a narc data function
This function makes modifications directly to an excel sheet. What I
still need to add is a restore function for specific rows because
sometimes it is over zealous in deleting. I also still need to address \
sometimes being addresses and other times being a city/state/zip. We
can't delete everything if Moorpark comes after the / because there is
also a Moorpark Street.